better floating point data types for zoom

+++ data/global_data.h
@@ -15,21 +15,21 @@ typedef struct {
 	int width;		// width in pixels
 	int height;		// height in pixels
 	int number_of_iterations;
-	double c_real;		// real aspect of c
-	double c_imag;		// imaginary aspect of c
-	double min_real;
-	double max_real;
-	double min_imag;
-	double max_imag;
+	long double c_real;		// real aspect of c
+	long double c_imag;		// imaginary aspect of c
+	long double min_real;
+	long double max_real;
+	long double min_imag;
+	long double max_imag;
 
-	double step_real;	// real difference between two columns
-	double step_imag;	// imag difference between two rows
+	long double step_real;	// real difference between two columns
+	long double step_imag;	// imag difference between two rows
 
 	int real_chunk_size;
 	int imag_chunk_size;
 
-	double current_real;
-	double current_imag;
+	long double current_real;
+	long double current_imag;
 
 	int current_chunk_id;
 	_Bool first_compute;
